Check whether a configuration entry with this key exists before accessing it
This prevents undefined index errors when accessing a configuration property which is not present in the configuration file. As an example if we want to take the DB persistence for the OAuth configuration and remove the token_persistence_path as proposed in the official documentation it will print an undefined index error. This will be resolved with this additional isset(...) check.