Allow configurable oauth logger
The oauth logger currently logs to the oauth directory in a hardcoded
path and filename. Allow it to be logged in any way by providing an
optional fully qualified classname to the user implementation of the
`ZohoOauthLoggerInterface`. If none is provided, use the default logger
which logs in exactly the same way as before.